Add event telemetry, curve mode, and airtime-aware scheduling
- Add MODE_EVENT: APRS beacons + burst event packets on flight state
  transitions (burnout, apogee, pyro fire/fail, deploy detect, landed,
  flight report). Events repeat N times for reliability.
- Add MODE_CURVE: max-rate barometric telemetry (pressure, altitude,
  velocity, state, decisecond timestamps) for post-flight curve
  reconstruction. No GPS, no APRS — just raw data as fast as possible.
- Add LoRa airtime calculator (lora_airtime.h/cpp) using Semtech
  SX1276 formula. All modes now dynamically compute minimum transmit
  intervals from SF/BW/CR instead of using static constants.
- Dead-air padding scales proportionally with airtime (20% for standard
  packets, 15% events, 10% curve) with floor values, so SF12 gets
  adequate spacing automatically.
- Extend $PYRO flags to 16-bit for altimeter event detection
  (drogue/main deploy OK/fail, pyro failure). Backward compatible
  with 2-char hex parsing.
- Add event_detector module: watches altimeter state/flag transitions,
  queues FlightEvent structs with flight time, altitude, velocity, GPS.
- Flight report event carries max altitude, max velocity, and error
  summary bitfield.

+++ src/altimeter_rx.h
@@ -4,6 +4,12 @@
 
 // All fields from the $PYRO NMEA sentence.
 // Values are integer, metric, matching the serial protocol exactly.
+//
+// The flags field is 16-bit to support extended event flags from the
+// altimeter's deployment/failure detection logic. Bits 0-5 are the
+// original flags; bits 6-11 are extended flags (EFLAG_*).
+// If the altimeter sends 2-char hex (old firmware), only bits 0-7 are
+// populated. If it sends 4-char hex (new firmware), all 16 bits are used.
 typedef struct {
     uint16_t seq;           // Sequence number (0–65535)
     uint8_t  state;         // 0=PAD_IDLE, 1=ASCENT, 2=DESCENT, 3=LANDED
@@ -13,7 +19,7 @@ typedef struct {
     int32_t  max_alt_cm;    // Max altitude this flight, cm
     int32_t  press_pa;      // Filtered pressure in pascals
     uint32_t flight_time_ms;// Flight time in ms since launch (0 on pad)
-    uint8_t  flags;         // Bitfield: p1_cont, p2_cont, p1_fired, p2_fired, armed, apogee
+    uint16_t flags;         // Bitfield: bits 0-5 original, bits 6-11 extended (EFLAG_*)
     uint16_t p1_adc;        // Pyro 1 raw ADC (0–4095)
     uint16_t p2_adc;        // Pyro 2 raw ADC (0–4095)
     uint16_t batt_adc;      // Battery voltage raw ADC (0 until implemented)
